big refactoring into logical classes. Should be easier to get an overview of the code now. I needed this to save time
